+++ app/src/main/java/com/example/internetshop/presentation/viewModel/ToolBarViewModel.kt
@@ -1,8 +1,14 @@
 package com.example.internetshop.presentation.viewModel
 
 import androidx.lifecycle.ViewModel
+import com.example.internetshop.R
 import javax.inject.Inject
 
 class ToolBarViewModel @Inject constructor(): ViewModel() {
-    val topLevelDestinations = mutableSetOf<Int>()
+    val topLevelDestinations = setOf(
+        R.id.authenticationFragment,
+        R.id.categoriesFragment,
+        R.id.favoriteListFragment,
+        R.id.cartFragment
+    )
 }
